requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
colorama==0.4.6
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
mysqlclient==2.1.1
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
protobuf==3.20.3
PyJWT==2.6.0
PyMySQL==1.0.3
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
